Input controls.

+++ payola/web/client/src/main/scala/cz/payola/web/client/views/entity/analysis/AnalysisOverviewView.scala
@@ -4,7 +4,7 @@ import cz.payola.web.client.presenters.components.AnalysisControls
 import cz.payola.web.client.views.ComposedView
 import cz.payola.common.entities.Analysis
 
-class AnalysisOverviewView(analysis: Analysis, timeoutSeconds: Long) extends ComposedView
+class AnalysisOverviewView(analysis: Analysis, timeoutSeconds: Int) extends ComposedView
 {
     val controls = new AnalysisControls(timeoutSeconds)
